World space UI button can plant in pots.
I made a prefab for putting UI in world space.

I made a button that havers above the pot that letes player plant. I'm switching the controls to have the mouse be the way players interact with game obects rather than a bunch of key press.

keyboard control is for player movement and camera control.

I'll be adding status bars to the world space UI.